refactor(common): consolidate duplicated OpenAI types into openaiTypes
OpenAIChatCompletionParams and OpenAIChatCompletionResponse were
defined in two places — OpenAI2GeminiConverter.ts and
OpenAI2AnthropicConverter.ts — with the Gemini definition being a
strict subset of the Anthropic one (missing max_tokens, temperature,
top_p, stop, stream).

The duplication was introduced when the Gemini converter was added in
#3069, where the author copy-pasted the interface from the older
Anthropic converter (#570) and dropped the optional fields. This
fragile setup got worse in #3241, which had to mirror the same JSDoc
into both files.

This refactor:
- Extracts both interfaces to a new shared module
  packages/desktop/src/common/api/openaiTypes.ts, with the union of
  fields (Anthropic-shaped; all extra fields are optional).
- Updates both converter files to import the shared types.
- Updates AnthropicRotatingClient and GeminiRotatingClient to import
  from the new location.
- Updates affected test files (OpenAI2GeminiConverter.test.ts,
  ProtocolConverter.test.ts) to import from the new location.

No behavior change:
- The Gemini converter reads only model, messages, tools (verified by
  grep). The extra fields in the shared interface (max_tokens,
  temperature, top_p, stop, stream) are simply not consulted.
- The Anthropic converter reads the same set of fields it always did.

Depends on #3241 (which establishes the [key: string]: unknown
index signature in the shared interface).

Closes #3245
