+++ packages/desktop/src/common/api/OpenAI2GeminiConverter.ts
@@ -5,72 +5,7 @@
  */
 
 import type { ProtocolConverter, ConverterConfig } from './ProtocolConverter';
-
-/**
- * OpenAI ChatCompletion-style params accepted by all rotating clients.
- *
- * Fields explicitly handled by the converters: `model`, `messages`, `tools`,
- * `tool_choice`. Any other OpenAI SDK field (e.g. `service_tier`,
- * `temperature`, `top_p`, `frequency_penalty`, `seed`, `logit_bias`,
- * `parallel_tool_calls`, etc.) is accepted via the index signature and
- * forwarded as-is to the OpenAI SDK. Non-OpenAI providers (Gemini,
- * Anthropic) silently drop these fields — pass-through is a no-op for them.
- *
- * The index signature is typed `unknown` (not `any`) per project lint
- * rules; callers can still spread a typed OpenAI object safely.
- */
-export interface OpenAIChatCompletionParams {
-  model: string;
-  messages: Array<{
-    role: string;
-    content:
-      | string
-      | Array<{
-          type: string;
-          text?: string;
-          image_url?: { url: string; detail?: string };
-        }>;
-  }>;
-  tools?: Array<{
-    type: 'function';
-    function: {
-      name: string;
-      description?: string;
-      parameters?: unknown;
-    };
-  }>;
-  tool_choice?: 'auto' | 'none' | { type: 'function'; function: { name: string } };
-  /**
-   * Pass-through for any additional OpenAI SDK parameters not enumerated
-   * above. Forwarded to the OpenAI SDK by `OpenAIRotatingClient`; ignored
-   * by Gemini/Anthropic converters.
-   */
-  [key: string]: unknown;
-}
-
-export interface OpenAIChatCompletionResponse {
-  id: string;
-  object: string;
-  created: number;
-  model: string;
-  choices: Array<{
-    index: number;
-    message: {
-      role: string;
-      content: string;
-      images?: Array<{
-        type: 'image_url';
-        image_url: { url: string };
-      }>;
-    };
-    finish_reason: string;
-  }>;
-  usage?: {
-    prompt_tokens: number;
-    completion_tokens: number;
-    total_tokens: number;
-  };
-}
+import type { OpenAIChatCompletionParams, OpenAIChatCompletionResponse } from './openaiTypes';
 
 export interface GeminiRequest {
   model: string;
